refactor: break out components and types
