+++ frontend/bacmet/app/search/components/fieldset.tsx
@@ -1,0 +1,5 @@
+export const FieldSet = ({children}: {children: React.ReactNode}) => {
+  return (
+    <fieldset className="form-group pt-3 pb-2">{children}</fieldset>
+  )
+}
