+++ frontend/bacmet/app/search/components/text-field.tsx
@@ -1,0 +1,19 @@
+import {useId} from "react";
+import {Field} from "../types";
+
+export const TextField = ({field}: {field: Field<unknown>}) => {
+  const fieldId = useId();
+  return (
+    <>
+      <label className="form-label" htmlFor={fieldId}>{field.label}:</label>
+      <input
+        className="form-control"
+        type="text"
+        name={field.name}
+        defaultValue={field.value !== undefined ? field.value + "" : undefined}
+        id={fieldId}
+        placeholder={field.placeholder}
+      />
+    </>
+  )
+}
