+++ frontend/bacmet/app/search/page.tsx
@@ -1,166 +1,13 @@
 'use client'
-import {useCallback, useEffect, useState, useId, FormEventHandler, Suspense} from "react";
+import {useCallback, useEffect, useState, FormEventHandler, Suspense} from "react";
 import {useConfig} from "../../contexts/config";
-import {useSearchParams, useRouter, usePathname} from 'next/navigation'
-
-type FieldValue<T=unknown> = {
-  value: T;
-  label: string;
-}
-
-type Field<T=unknown> = {
-  label: string;
-  name: string;
-  value?: T;
-  values: FieldValue<T>[];
-  placeholder?: string;
-}
-
-type SearchParams = {
-  chemicalClasses: FieldValue<string>[];
-  compounds: FieldValue<string>[];
-}
-
-type Compound = {
-  compound_name: string;
-}
-
-type Validated = {
-  gene_name: string;
-  bacmet_id: string;
-  code_for: string;
-  family: string;
-  organism: string;
-  location: string;
-  compounds: Compound[];
-  description: string;
-  length_aa: string;
-  reference: string;
-}
-
-type Predicted = {
-  gene_name: string;
-  protein_accession_uniprot: string;
-  organism: string;
-  compounds: Compound[];
-}
-
-type Meta = {
-  totalRecords: number;
-  page: number;
-  count: number;
-}
-
-type Link = {
-  href: string;
-  rel: string;
-}
-
-type Result<T> = {
-  items: T[];
-  _meta: Meta[];
-  _links: Link[];
-}
-
-type ValidatedResult = {
-  type: "validated"
-} & Result<Validated>
-
-type PredictedResult = {
-  type: "predicted"
-} & Result<Predicted>
-
-const SelectField = ({field}: {field: Field<unknown>,}) => {
-  const fieldId = useId();
-  return (
-    <>
-      <label className="form-label" htmlFor={fieldId}>{ field.label }:</label>
-      <select className="form-control" name={field.name} id={fieldId} defaultValue={field.value + ""}>
-        {field.values.map(value => (
-          <option key={value.value + ""} value={value.value + ""}>{value.label}</option>
-        ))}
-      </select>
-    </>
-  )
-}
-
-const RadioSelectField = ({field}: {field: Field<unknown>}) => {
-  const fieldId = useId();
-  return (
-    <>
-      <legend>{ field.label }:</legend>
-      {field.values.map(value => {
-        const valueId = `${fieldId}-${value.value}`;
-        return (
-          <div key={value.value + ""} className="form-check form-check-inline">
-            <input
-              className="form-check-input"
-              type="radio"
-              name={field.name}
-              defaultChecked={value.value==field.value} 
-              value={value.value + ""}
-              id={valueId}
-              />
-            <label
-              className="form-check-label"
-              htmlFor={valueId}
-            >{ value.label }</label>
-          </div>
-        )
-      })}
-    </>
-  )
-}
-
-const TextField = ({field}: {field: Field<unknown>}) => {
-  const fieldId = useId();
-  return (
-    <>
-      <label className="form-label" htmlFor={fieldId}>{field.label}:</label>
-      <input
-        className="form-control"
-        type="text"
-        name={field.name}
-        defaultValue={field.value !== undefined ? field.value + "" : undefined}
-        id={fieldId}
-        placeholder={field.placeholder}
-      />
-    </>
-  )
-}
-
-const FieldSet = ({children}: {children: React.ReactNode}) => {
-  return (
-    <fieldset className="form-group pt-3 pb-2">{children}</fieldset>
-  )
-}
-
-const Pagination = (
-  {pages, currentPage, label, onNavigate}: {pages: Link[], currentPage?: string, label?: string, onNavigate: (page: Link) => void}
-) => {
-  const maxPages = 5;
-  const first: Link = pages[0]
-  const last: Link = pages[pages.length - 1]
-  const currentIndex = pages.findIndex((link) => link.href === currentPage);
-  const firstIndex = Math.ceil(Math.max(1, currentIndex - maxPages / 2));
-  const lastIndex = Math.min(pages.length - 1, firstIndex + maxPages);
-  const pageList = [
-    ...(first === undefined ? [] : [{...first, rel: "First"}]),
-    ...pages.slice(firstIndex, lastIndex),
-    ...(last === undefined ? [] : [{...last, rel: `Last (${last.rel})`}])
-  ]
-  return (
-    <nav aria-label={label || "Pagination"}>
-      <ul className="pagination">
-        {pageList.map((page, index) => (
-          <li key={index} className={`page-item ${ page.href === currentPage ? "active" : "" }`}>
-            <span className="page-link" onClick={() => onNavigate(page)}>{ page.rel }</span>
-          </li>
-        ))}
-      </ul>
-    </nav>
-  )
-}
+import {useSearchParams, useRouter, usePathname} from "next/navigation";
+import {SearchParams, ValidatedResult, PredictedResult, Field, Link} from "./types";
+import {FieldSet} from "./components/fieldset";
+import {Pagination} from "./components/pagination";
+import {RadioSelectField} from "./components/radio-select-field";
+import {SelectField} from "./components/select-field";
+import {TextField} from "./components/text-field";
 
 const SearchBase = (
   {values}: {
